+++ Sources/ZoomImageViewer/_Extensions-internal/ContainerCornerOffsetIfAvailable.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 extension View {
-    nonisolated public func containerCornerOffsetIfAvailable(_ edges: Edge.Set, sizeToFit: Bool = false) -> some View {
+    nonisolated func containerCornerOffsetIfAvailable(_ edges: Edge.Set, sizeToFit: Bool = false) -> some View {
         ifAvailable {
             #if compiler(>=6.2)
             if #available(iOS 26, *) {